+++ src/prefect/core/flow.py
@@ -33,7 +33,8 @@ from prefect.core.edge import Edge
 from prefect.core.task import Parameter, Task
 from prefect.engine.result import NoResult
 from prefect.engine.result_handlers import ResultHandler
-from prefect.environments import Environment
+from prefect.environments import CloudEnvironment, Environment
+from prefect.environments.storage import Storage
 from prefect.utilities import logging
 from prefect.utilities.notifications import callback_factory
 from prefect.utilities.serialization import to_qualified_name
@@ -106,7 +107,9 @@ class Flow:
         - name (str): The name of the flow. Cannot be `None` or an empty string
         - schedule (prefect.schedules.Schedule, optional): A default schedule for the flow
         - environment (prefect.environments.Environment, optional): The environment
-            type that the flow should be run in. If None, a LocalEnvironment will be created.
+           that the flow should be run in. If `None`, a `CloudEnvironment` will be created.
+        - storage (prefect.environments.storage.Storage, optional): The unit of storage
+            that the flow will be written into.
         - tasks ([Task], optional): If provided, a list of tasks that will initialize the flow
         - edges ([Edge], optional): A list of edges between tasks
         - reference_tasks ([Task], optional): A list of tasks which determine the final
@@ -136,6 +139,7 @@ class Flow:
         name: str,
         schedule: prefect.schedules.Schedule = None,
         environment: Environment = None,
+        storage: Storage = None,
         tasks: Iterable[Task] = None,
         edges: Iterable[Edge] = None,
         reference_tasks: Iterable[Task] = None,
@@ -146,9 +150,6 @@ class Flow:
     ):
         self._cache = {}  # type: dict
 
-        # set random id
-        self.id = str(uuid.uuid4())
-
         self.logger = logging.get_logger("Flow")
 
         if not name:
@@ -156,7 +157,8 @@ class Flow:
 
         self.name = name
         self.schedule = schedule
-        self.environment = environment or prefect.environments.LocalEnvironment()
+        self.environment = environment or prefect.environments.CloudEnvironment()
+        self.storage = storage
         self.result_handler = (
             result_handler or prefect.engine.get_default_result_handler_class()()
         )
@@ -210,8 +212,6 @@ class Flow:
         new = copy.copy(self)
         # create a new cache
         new._cache = dict()
-        # create new id
-        new.id = str(uuid.uuid4())
         new.tasks = self.tasks.copy()
         new.edges = self.edges.copy()
         new.set_reference_tasks(self._reference_tasks)
@@ -308,30 +308,6 @@ class Flow:
 
         if validate:
             self.validate()
-
-    @property
-    def id(self) -> str:
-        return self._id
-
-    @id.setter
-    def id(self, value: str) -> None:
-        """
-        Args:
-            - value (str): a UUID-formatted string
-        """
-        try:
-            uuid.UUID(value)
-        except Exception:
-            raise ValueError("Badly formatted UUID string: {}".format(value))
-        self._id = value
-
-    @property  # type: ignore
-    @cache
-    def task_ids(self) -> Dict[str, Task]:
-        """
-        Returns a dictionary of {task_id: Task} pairs.
-        """
-        return {task.id: task for task in self.tasks}
 
     # Context Manager ----------------------------------------------------------
 
@@ -445,7 +421,7 @@ class Flow:
                 "Tasks must be Task instances (received {})".format(type(task))
             )
         elif task not in self.tasks:
-            if task.slug and task.slug in [t.slug for t in self.tasks]:
+            if task.slug and any(task.slug == t.slug for t in self.tasks):
                 raise ValueError(
                     'A task with the slug "{}" already exists in this '
                     "flow.".format(task.slug)
@@ -1116,15 +1092,6 @@ class Flow:
 
     # Building / Serialization ----------------------------------------------------
 
-    def to_environment_file(self, path: str) -> None:
-        """
-        Serializes the flow as an environment file.
-
-        Args:
-            - path (str): the path of the environment file to create
-        """
-        self.environment.build(self).to_file(path)
-
     def serialize(self, build: bool = False) -> dict:
         """
         Creates a serialized representation of the flow.
@@ -1139,18 +1106,16 @@ class Flow:
 
         self.validate()
         schema = prefect.serialization.flow.FlowSchema
-        serialized = schema(exclude=["environment"]).dump(self)
+        serialized = schema(exclude=["storage"]).dump(self)
 
         if build:
-            environment = self.environment.build(
-                flow=self
-            )  # type: Optional[Environment]
+            if not self.storage:
+                raise ValueError("This flow has no storage to build")
+            storage = self.storage.build(flow=self)  # type: Optional[Storage]
         else:
-            environment = self.environment
+            storage = self.storage
 
-        serialized.update(
-            schema(only=["environment"]).dump({"environment": environment})
-        )
+        serialized.update(schema(only=["storage"]).dump({"storage": storage}))
 
         return serialized
 
@@ -1215,7 +1180,13 @@ class Flow:
         ids = {}
         for t in tasks:
             serialized = t.serialize()
-            del serialized["id"]  # remove the ID since it is unique but random
+            try:
+                uuid.UUID(serialized["slug"])
+                del serialized[
+                    "slug"
+                ]  # remove the slug since it is randomly generated UUID
+            except ValueError:
+                pass  # don't remove slug because it was set manually
             ids[t] = _hash(json.dumps(serialized, sort_keys=True))
 
         if _debug_steps: